1.完善回答详情底部操作栏 2.修正通知中心查看已读消息会删除项的bug

+++ app/src/main/java/com/twt/service/wenjin/ui/notification/readlist/NotificationFragment.java
@@ -191,6 +191,7 @@ public class NotificationFragment extends BaseFragment implements NotificationVi
 
     @Override
     public void deleteItem(int argPosition) {
+        if(type == 1 ){return;}
         NotificationItem item = mAdapter.getItems(argPosition);
         mPresenter.markNotificationAsRead(item.notification_id);
         mAdapter.deleteItem(argPosition);
